container/webapp/requirements.txt
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-compute==1.8.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
prettytable==3.5.0
proto-plus==1.22.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.1
rsa==4.9
six==1.16.0
tcping==0.1.1rc1
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.11.0
